refactor(tests): extract shared lib for resilience + multi-pod stacks
Both Docker-Compose-backed suites had separately-maintained copies of:
  - poll-until.ts (1:1 duplicate; multi-pod's doc was slightly nicer)
  - the fetch/header-building logic (Origin/Bearer/Cookie) — only the
    base URL differed (constant in resilience, per-pod in multi-pod)

Centralizes both in tests/shared/lib/. Suite-local files become thin
wrappers/re-exports so scenario imports don't move.

Net: 78 lines deleted, one canonical implementation of "what does
Better Auth's CSRF guard expect from a test client" — no more risk of
the two stacks drifting on header shape.

Out of scope here: merging the docker-compose.yml files. Per earlier
review the topologies (toxiproxy chaos vs healthy multi-pod) are part
of what each suite tests — they shouldn't merge.
